requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
aiohttp-retry==2.4.5
altair==4.1.0
anyio==3.3.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atpublic==2.3
Babel==2.9.1
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
base58==2.1.0
beautifulsoup4==4.9.3
blinker==1.4
blis==0.4.1
boto3==1.19.1
botocore==1.22.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
ci-info==0.2.0
clang==5.0
click==7.1.2
commonmark==0.9.1
configobj==5.0.6
configparser==5.0.2
cycler==0.10.0
cymem==2.0.6
Cython==0.29.14
dictdiffer==0.9.0
diskcache==5.2.1
distlib==0.3.3
distro==1.6.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.7.4
etelemetry==0.2.2
filelock==3.0.12
fitz==0.0.1.dev2
Flask==2.0.1
flatbuffers==1.12
flatten-dict==0.4.2
flufl.lock==3.2
fsspec==2021.9.0
ftfy==5.9
funcy==1.16
future==0.18.2
gast==0.2.2
gensim==3.8.3
gitdb==4.0.7
GitPython==3.1.24
google-api-core==2.0.1
google-auth==1.34.0
google-auth-oauthlib==0.4.4
google-cloud-texttospeech==2.5.2
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grandalf==0.6
grpcio==1.40.0
h5py==3.1.0
httplib2==0.19.1
huggingface-hub==0.0.19
idna==3.2
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==2.0.1
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jupyter-server==1.10.1
jupyterlab==3.1.0
jupyterlab-server==2.6.1
keras==2.6.0
Keras-Applications==1.0.8
keras-nightly==2.7.0.dev2021091807
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
laserembeddings==1.1.1
libclang==11.1.0
lxml==4.6.3
mailchecker==4.0.12
Markdown==3.3.4
matplotlib==3.4.2
multidict==5.1.0
murmurhash==1.0.5
nanotime==0.5.2
nbclassic==0.3.1
networkx==2.6.2
nibabel==3.2.1
nipype==1.6.1
nltk==3.6.2
numpy==1.19.5
numpyencoder==0.3.0
oauthlib==3.1.1
opt-einsum==3.3.0
pandas==1.3.1
pathlib==1.0.1
pathspec==0.8.1
pathy==0.6.0
pdfminer.six==20200517
pdfplumber==0.5.28
phonenumbers==8.12.33
Pillow==8.3.1
plac==0.9.6
platformdirs==2.4.0
pluggy==1.0.0
ply==3.11
preshed==3.0.6
proto-plus==1.19.0
protobuf==3.17.3
prov==2.0.0
psutil==5.8.0
py==1.10.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pydantic==1.8.2
pydeck==0.7.0
pydot==1.4.2
pygit2==1.6.1
pygtrie==2.4.2
PyMuPDF==1.18.15
pyparsing==2.4.7
PyPDF2==1.26.0
pyrsistent==0.18.0
python-benedict==0.24.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytokenizations==0.8.4
pytz==2021.1
pywin32==301
pywinpty==1.1.3
pyxnat==1.4
PyYAML==5.4.1
rdflib==6.0.0
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rich==10.11.0
rouge==1.0.1
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.5.0
sacremoses==0.0.35
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
shortuuid==1.0.1
shtab==1.4.1
simplejson==3.17.3
sklearn==0.0
smart-open==5.1.0
smmap==4.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
spacy==3.1.3
spacy-alignments==0.8.3
spacy-legacy==3.0.8
spacy-transformers==1.0.6
speedcopy==2.1.0
srsly==2.4.2
streamlit==1.0.0
subword-nmt==0.3.7
tabulate==0.8.9
tb-nightly==2.7.0a20210917
tensorboard==2.0.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
terminado==0.10.1
text-unidecode==1.3
textblob==0.15.3
tf-estimator-nightly==2.7.0.dev2021091808
thinc==8.0.11
thinc-gpu-ops==0.0.4
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
toolz==0.11.1
torch==1.9.0+cu102
torchaudio==0.9.0
torchcontrib==0.0.2
torchvision==0.10.0+cu102
tox==3.24.4
tqdm==4.61.2
traits==6.2.0
transformers==4.11.3
transliterate==1.10.2
typer==0.3.2
tzdata==2021.2.post0
tzlocal==3.0
urllib3==1.26.6
validators==0.18.2
virtualenv==20.8.1
voluptuous==0.12.2
Wand==0.6.6
wasabi==0.8.2
watchdog==2.1.6
webencodings==0.5.1
websocket-client==1.1.0
Werkzeug==2.0.1
win-unicode-console==0.5
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
yellowbrick==1.3.post1
zc.lockfile==2.0

